+++ lib/debug/repl.hpp
@@ -11,7 +11,7 @@
 #include <io/hardware_gpio.hpp>
 #include <io/hardware_i2c.hpp>
 #include <io/hardware_spi.hpp>
-#include <io/uart.hpp>
+#include <io/hardware_uart.hpp>
 
 namespace hyped::debug {
